Compare int value to EOF
While we indeed are acquiring a char and comparing it, both value returned by getc, fgetc and EOF constant are int, not char, so we should be using them as int instead of char. Using the wrong type may also break the encryption and decryption, as char supports a way limited set of codes.